+++ android/src/main/java/com/tenforwardconsulting/bgloc/DistanceFilterLocationProvider.java
@@ -167,7 +167,7 @@ public class DistanceFilterLocationProvider extends AbstractLocationProvider imp
     }
 
     /**
-    * Translates a number representing desired accuracy of GeoLocation system from set [0, 10, 100, 1000].
+    * Translates a number representing desired accuracy of Geolocation system from set [0, 10, 100, 1000].
     * 0:  most aggressive, most accurate, worst battery drain
     * 1000:  least aggressive, least accurate, best for battery.
     */